port_to_agent_sdk/requirements.txt
anthropic
anyio
claude-agent-sdk
python-dotenv
